requirements.txt
beautifulsoup4
openai
python-dotenv
requests
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client